Exercise 15 completed using forEach
That was the first way to solve the problem. Asking chatGPT for correction it suggested using map
to create the second array, but didn't gave me the solution. So I'm making this commit to save
my first version of this solution and will make another one using the method it proposed